chore: Refactor run_integration_tests.sh by adding stack language check
