+++ integration_tests/stacks/typescript/lambda-provided-stack.ts
@@ -3,7 +3,7 @@ import * as s3 from "aws-cdk-lib/aws-s3";
 import { LambdaRestApi, LogGroupLogDestination } from "aws-cdk-lib/aws-apigateway";
 import { LogGroup } from "aws-cdk-lib/aws-logs";
 import { Stack, StackProps, App } from "aws-cdk-lib";
-import { DatadogLambda } from "../../src/index";
+import { DatadogLambda } from "../../../src/index";
 
 export class ExampleStack extends Stack {
   constructor(scope: App, id: string, props?: StackProps) {
@@ -18,7 +18,6 @@ export class ExampleStack extends Stack {
       runtime: lambda.Runtime.PROVIDED_AL2,
       code: lambda.Code.fromBucket(s3Bucket, "fake-key-for-test"),
       handler: "handler.handler",
-      architecture: lambda.Architecture.ARM_64,
     });
 
     s3Bucket.grantRead(providedLambda);
@@ -44,6 +43,6 @@ export class ExampleStack extends Stack {
 
 const app = new App();
 const env = { account: "601427279990", region: "sa-east-1" };
-const stack = new ExampleStack(app, "lambda-provided-arm-stack", { env: env });
+const stack = new ExampleStack(app, "lambda-provided-stack", { env: env });
 console.log("Stack name: " + stack.stackName);
 app.synth();
